# api/swagger/swagger.yaml
swagger: '2.0'
info:
  description: Communication with Slack API
  version: '1'
  title: Slack API
host: localhost:10010
basePath: /
schemes:
  - https
paths:
...
  /sendMsg/schedule:
    x-swagger-router-controller: sendMsgSchedule
    post:
      operationId: sendMsgSchedule
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: false
          schema:
            properties:
              channel:
                type: string
              postAt:
                type: string
              text:
                type: string
      responses:
        '200':
          description: Definition generated from Swagger Inspector
          schema:
            type: string
  /sendMsg/schedule/oneButton/:
    x-swagger-router-controller: sendMsgScheduleOneButton
    post:
      operationId: sendMsgScheduleOneButton
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: false
          schema:
            properties:
              channel:
                type: string
              postAt:
                type: string
              text:
                type: string
              callbackId:
                type: string 
              textButton1:
                type: string
      responses:
        '200':
          description: Definition generated from Swagger Inspector
          schema:
            type: string
  /sendMsg/schedule/oneButton/confirm:
    x-swagger-router-controller: sendMsgScheduleOneButtonConfirm
    post:
      operationId: sendMsgScheduleOneButtonConfirm
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: false
          schema:
            properties:
              channel:
                type: string
              postAt:
                type: string
              text:
                type: string
              callbackId:
                type: string 
              textButton1:
                type: string
              textConfirmation1:
                type: string
      responses:
        '200':
          description: Definition generated from Swagger Inspector
          schema:
            type: string
  /sendMsg/schedule/twoButtons:
    x-swagger-router-controller: sendMsgScheduleTwoButtons
    post:
      operationId: sendMsgScheduleTwoButtons
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: false
          schema:
            properties:
              channel:
                type: string
              postAt:
                type: string
              text:
                type: string
              callbackId:
                type: string 
              textButton1:
                type: string
              textButton2:
                type: string
      responses:
        '200':
          description: Definition generated from Swagger Inspector
          schema:
            type: string
  /sendMsg/schedule/twoButtons/confirm:
    x-swagger-router-controller: sendMsgScheduleTwoButtonsConfirm
    post:
      operationId: sendMsgScheduleTwoButtonsConfirm
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: false
          schema:
            properties:
              channel:
                type: string
              postAt:
                type: string
              text:
                type: string
              callbackId:
                type: string 
              textButton1:
                type: string
              textButton2:
                type: string
              textConfirmation1:
                type: string
              textConfirmation2:
                type: string
      responses:
        '200':
          description: Definition generated from Swagger Inspector
          schema:
            type: string
...
  /startDialog/oneTextElement:
    x-swagger-router-controller: startDialog
    post:
      operationId: startDialogOneTextElement
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: false
          schema:
            properties:
              triggerid:
                type: string
              callbackId:
                type: string
              title:
                type: string
              label1:
                type: string
              name1:
                type: string
              placeholder1:
                type: string
      responses:
        '200':
          description: Definition generated from Swagger Inspector
          schema:
            type: string
  /startDialog/twoTextElements:
    x-swagger-router-controller: startDialog
    post:
      operationId: startDialogTwoTextElements
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: false
          schema:
            properties:
              triggerid:
                type: string
              callbackId:
                type: string
              title:
                type: string
              label1:
                type: string
              name1:
                type: string
              placeholder1:
                type: string
              label2:
                type: string
              name2:
                type: string
              placeholder2:
                type: string
      responses:
        '200':
          description: Definition generated from Swagger Inspector
          schema:
            type: string
  /startDialog/selectMenus:
    x-swagger-router-controller: startDialog
    post:
      operationId: startDialogSelectMenus
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: false
          schema:
            properties:
              triggerid:
                type: string
              callbackId:
                type: string
              title:
                type: string
              label1:
                type: string
              name1:
                type: string
      responses:
        '200':
          description: Definition generated from Swagger Inspector
          schema:
            type: string
...